backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
typing-extensions>=4.0.0
sqlalchemy>=1.4.0
python-multipart>=0.0.5
pandas
immudb-py
pytest
httpx